Adding support to load NRRD volume files.

+++ ultraliser/algorithms/DualMarchingCubes.h
@@ -50,7 +50,7 @@ public:
      * @param isoValue
      */
     DualMarchingCubes(Volume* volume,
-                      const uint8_t isoValue = DEFAULT_ISO_VALUE,
+                      const uint64_t isoValue = DEFAULT_ISO_VALUE,
                       const bool &generateManifold = true);
 
     /**
@@ -198,7 +198,7 @@ private:
     /**
      * @brief _isoValue
      */
-    const uint8_t _isoValue;
+    const uint64_t _isoValue;
 
     /**
      * @brief _generateManifold
