+++ ultraliser/data/volumes/grids/Projection.h
@@ -39,11 +39,11 @@ namespace Ultraliser
  * @param colorMap
  */
 void saveColorMappedProjection(const std::string &prefix,
-                               const float* projection,
+                               const double* projection,
                                const uint64_t &projectionWidth,
                                const uint64_t &projectionHeight,
-                               const float& minValue,
-                               const float& maxValue,
+                               const double& minValue,
+                               const double& maxValue,
                                const std::string &colorMap);
 
 /**
@@ -56,11 +56,11 @@ void saveColorMappedProjection(const std::string &prefix,
  * @param maxValue
  */
 void saveColorMappedProjectionWithAllColorMaps(const std::string &prefix,
-                                               const float* projection,
+                                               const double *projection,
                                                const uint64_t &projectionWidth,
                                                const uint64_t &projectionHeight,
-                                               const float& minValue,
-                                               const float& maxValue);
+                                               const double &minValue,
+                                               const double &maxValue);
 
 }
